move all data_utils stuff to Event class
