+++ src/OSmOSE/data/base_item.py
@@ -10,7 +10,7 @@ from typing import TYPE_CHECKING, Generic, TypeVar
 import numpy as np
 
 from OSmOSE.data.base_file import BaseFile
-from OSmOSE.utils.data_utils import Event
+from OSmOSE.data.event import Event
 
 if TYPE_CHECKING:
     from pandas import Timestamp
